requirements.txt
selenium==3.141.0
Appium-Python-Client==0.52
urllib3<2.0.0
pywin32>=310
psutil>=5.9.0
pyinstaller==6.5.0 